requirements/base.txt
Django==2.2.3
django-filter==2.2.0
django-rest-swagger==2.1.2
djangorestframework==3.9.4
Markdown==3.1.1
python-decouple==3.1
PyYAML==5.1.1
requests==2.22.0
mysqlclient==1.4.2.post1
factory-boy==2.12.0
pandas==0.25.0